Add files via upload
A major appeal of Open Science is the capacity to harness collective intelligence and face big scientific challenges, unmet when researchers are working in isolation. This file is a list of such projects. We worked your way through the list as a group to discover the projects. Then, applied our scores (1-5 low-to-high score), using 10 of the Vienna Open Science Principles to rank the projects’ ‘openness’. Finally, added a few comments on what we appreciated and what we think could be improved in these projects.